requirements.txt
numpy>=1.20.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
tqdm>=4.62.0
pillow>=9.0.0
opencv-python>=4.5.0
scikit-image>=0.19.0
open3d>=0.16.0
scipy>=1.7.0
requests>=2.27.0
timm>=0.9.0  # For Vision Transformer support
transformers>=4.30.0  # For advanced model components
einops>=0.6.0  # For tensor manipulation in ViT
tensorboard>=2.13.0  # For training visualization